just got the pages working :D
