--- client/src/components/Landing.js
@@ -1,31 +1,0 @@
-import React from 'react';
-import '../styles/Landing.css'
-
-function Landing() {
-    return (
-    <div class="landingcontainer">
-        <h1>
-            ProPairs
-        </h1>
-        <h2>
-            Group cohesion is everything in projects. We get it. ProPairs evaluates users based on their personality type and generates a rating of how well they will work with one another. The website aims to help users find compatible project partners based on their personality traits.
-        </h2>
-        <button class="button-front" role="button">Get Started</button>
-    </div>
-)}
-
-//   <div class="yay">
-//     <h3>
-//         How it works
-//     </h3>
-//     <div class="infoContainer">
-//         <h2>
-//             Group cohesion is everything in projects. We get it. ProPairs evaluates users based on their personality type and generates a rating of how well they will work with one another. The website aims to help users find compatible project partners based on their personality traits.
-//             Group cohesion is everything in projects. We get it. ProPairs evaluates users based on their personality type and generates a rating of how well they will work with one another. The website aims to help users find compatible project partners based on their personality traits.
-//             Group cohesion is everything in projects. We get it. ProPairs evaluates users based on their personality type and generates a rating of how well they will work with one another. The website aims to help users find compatible project partners based on their personality traits.
-//             Group cohesion is everything in projects. We get it. ProPairs evaluates users based on their personality type and generates a rating of how well they will work with one another. The website aims to help users find compatible project partners based on their personality traits.
-//         </h2>
-//     </div>
-// </div>
-
-export default Landing
