+++ client/src/components/PortfolioContainer.js
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import NavTabs from './NavTabs';
-import Home from './pages/Home';
-import Profile from './pages/Profile';
-import MyTeams from './pages/MyTeams';
-import JoinTeam from './pages/JoinTeam';
+import Home from '../pages/Home';
+import Profile from '../pages/Profile';
+import MyTeams from '../pages/MyTeams';
+import JoinTeam from '../pages/JoinTeam';
 
 export default function PortfolioContainer() {
   const [currentPage, setCurrentPage] = useState('Home');
