+++ client/src/pages/JoinTeam.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+function joinTeam() {
+    return (
+        <div>
+            Join Tream
+        </div>
+    )
+}
+
+export default joinTeam;
