+++ client/src/pages/MyTeams.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+function myTeam() {
+    return (
+        <div>
+            MyTeam
+        </div>
+    )
+}
+
+export default myTeam;
