+++ client/src/pages/Profile.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Profile() {
+    return (
+        <div>
+            Profile
+        </div>
+    )
+}
+
+export default Profile;
